move workspace workflows into a separate file
